* switched to floats for time periods in hours, so you can set 30 mins as 0.5 hours

+++ JSONModel/JSONModelNetworking/JSONCache.h
@@ -18,21 +18,62 @@
 #import "JSONCacheFile.h"
 #import "JSONCacheResponse.h"
 
-extern int kNeverRevalidate;
-extern int kAlwaysRevalidate;
+//use these when setting the cache revalidation times
+extern float kNeverRevalidate;
+extern float kAlwaysRevalidate;
 
-extern int kNeverExpire;
-extern int kImmediatelyExpire;
+//use these when setting the cache expiration times
+extern float kNeverExpire;
+extern float kImmediatelyExpire;
 
+/**
+ * Cache engine for your JSON based API - it helps you easily get local cache,
+ * offline cache, and also get time or etag based server revalidation.
+ *
+ * You can use this class with <JSONHTTPClient> or separately with your own code.
+ *
+ * The basic idea behine JSONCache are the hard-expiration and soft-expiration times.
+ * You set hard-expiration (by setting **expirationTimeInHours** and **expirationTimeInHoursWhenOffline**
+ * so that objects in the cache are considered expired after the given amount of time
+ * but the amount of time could be different depending whether the user is online or offline
+ * (this is mostly applicable on the iPhone).
+ *
+ * Here's a short example how to set your <JSONHTTPClient> and <JSONAPI> to support 
+ * caching API responses for 30 minutes, but keep all objects practicly forever cached
+ * when the user is not connected to Internet (aka providing offline mode for your app)
+ * <pre>
+ * [JSONCache sharedCache].expirationTimeInHours = 0.5;
+ * [JSONCache sharedCache].expirationTimeInHoursWhenOffline = kNeverExpire;
+ * [JSONHTTPClient setIsUsingJSONCache: YES];
+ *
+ * [[JSONCache sharedCache] loadCacheFromDisc];
+ *
+ * [JSONAPI postWithPath:@"/getObjects" andParams:nil completion:... 
+ * ... here use JSONAPI or JSONHTTPClient normally ...
+ * ... the cache is already working with them with the config you've set ...
+ * </pre>
+ *
+ * NB: If you are persisting your cache between app sessions (the common case) for apps
+ * with offline mode, do not forget after configuring the cache, to load the cached files
+ * from the device's disc: [[JSONCache sharedCache] loadCacheFromDisc]
+ * 
+ * If you want to check what are the cached objects in the cache right now: 
+ * <pre>
+ * NSLog(@"Cache: %@", [JSONCache sharedCache]);
+ * </pre>
+ * Setting your cache to automatically use server re-validation works in a similar fashion
+ * you just need to use **revalidateCacheFromServerAfterTimeInHours** or **revalidateCacheViaETagAfterTimeInHours**
+ * but for this to work you also will need to adjust your backend logic.
+ */
 @interface JSONCache : NSObject
 
 @property (assign, nonatomic, readonly) BOOL isOnline;
 
-@property (assign, nonatomic) int expirationTimeInHours;
-@property (assign, nonatomic) int expirationTimeInHoursWhenOffline;
+@property (assign, nonatomic) float expirationTimeInHours;
+@property (assign, nonatomic) float expirationTimeInHoursWhenOffline;
 
-@property (assign, nonatomic) int revalidateCacheFromServerAfterTimeInHours;
-@property (assign, nonatomic) int revalidateCacheViaETagAfterTimeInHours;
+@property (assign, nonatomic) float revalidateCacheFromServerAfterTimeInHours;
+@property (assign, nonatomic) float revalidateCacheViaETagAfterTimeInHours;
 
 @property (assign, nonatomic) BOOL isUsingXdHTTPHeaderNames;
